+++ src/main/java/uk/gov/hmcts/sptribs/edgecase/event/cic/CicSubmitCaseEvent.java
@@ -7,14 +7,14 @@ import uk.gov.hmcts.ccd.sdk.api.CCDConfig;
 import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
 import uk.gov.hmcts.ccd.sdk.api.ConfigBuilder;
 import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
+import uk.gov.hmcts.sptribs.ciccase.model.CaseData;
+import uk.gov.hmcts.sptribs.ciccase.model.State;
+import uk.gov.hmcts.sptribs.ciccase.model.UserRole;
 import uk.gov.hmcts.sptribs.common.ccd.CcdCaseType;
 import uk.gov.hmcts.sptribs.common.config.AppsConfig;
-import uk.gov.hmcts.sptribs.edgecase.model.CaseData;
-import uk.gov.hmcts.sptribs.edgecase.model.State;
-import uk.gov.hmcts.sptribs.edgecase.model.UserRole;
 import uk.gov.hmcts.sptribs.util.AppsUtil;
 
-import static uk.gov.hmcts.sptribs.edgecase.model.UserRole.CITIZEN;
+import static uk.gov.hmcts.sptribs.ciccase.model.UserRole.CITIZEN;
 import static uk.gov.hmcts.sptribs.edgecase.model.access.Permissions.CREATE_READ_UPDATE;
 
 @Component
@@ -30,7 +30,7 @@ public class CicSubmitCaseEvent implements CCDConfig<CaseData, State, UserRole> 
         configBuilder
             .event(AppsUtil.getExactAppsDetailsByCaseType(appsConfig, CcdCaseType.CIC.getCaseTypeName()).getEventIds()
                        .getSubmitEvent())
-            .forStates(State.DRAFT)
+            .forStates(State.Draft)
             .name("Submit case (cic)")
             .description("Applicant confirms SOT (cic)")
             .retries(120, 120)
@@ -39,7 +39,7 @@ public class CicSubmitCaseEvent implements CCDConfig<CaseData, State, UserRole> 
     }
 
     public AboutToStartOrSubmitResponse<CaseData, State> aboutToSubmit(CaseDetails<CaseData, State> details,
-                                                                       CaseDetails<CaseData, State> beforeDetails) {
+                                                                           CaseDetails<CaseData, State> beforeDetails) {
 
         //TODO logic needs to be updated separately as per edge-case application requirement
         return AboutToStartOrSubmitResponse.<CaseData, State>builder()
