+++ src/main/java/uk/gov/hmcts/sptribs/edgecase/model/access/DefaultAccess.java
@@ -6,8 +6,8 @@ import uk.gov.hmcts.ccd.sdk.api.HasAccessControl;
 import uk.gov.hmcts.ccd.sdk.api.HasRole;
 import uk.gov.hmcts.ccd.sdk.api.Permission;
 
-import static uk.gov.hmcts.sptribs.edgecase.model.UserRole.CITIZEN;
-import static uk.gov.hmcts.sptribs.edgecase.model.UserRole.CREATOR;
+import static uk.gov.hmcts.sptribs.ciccase.model.UserRole.CITIZEN;
+import static uk.gov.hmcts.sptribs.ciccase.model.UserRole.CREATOR;
 
 public class DefaultAccess implements HasAccessControl {
     @Override
