--- src/main/java/uk/gov/hmcts/sptribs/edgecase/search/CaseFieldsConstants.java
@@ -1,11 +1,0 @@
-package uk.gov.hmcts.sptribs.edgecase.search;
-
-public final class CaseFieldsConstants {
-
-    public static final String SUBJECT_FULL_NAME = "subjectFullName";
-    public static final String FULL_NAME = "FullName";
-
-    // required for Checkstyle
-    private CaseFieldsConstants() {
-    }
-}
